Made some changes to the way walking works
it's now physics based by assigning the player a velocity instead of just moving them forward. I tried to implement a different walking style but it ended up being more complicated and didn't work entirely due to an issue with changing direction and getting stuck until you let go of all the keys. I know it has to do with have two seperate coroutines, but because i managed to get the functionality out of the original code I've opted to go with that. I've left my other attempt in the file in case i want to revisit it.