Balcony Area, still a WIP
I want to try loading the StageFloor undernearth, slightly blurred and smaller so that you can see it from the balcony, but how I'm going to implement that is not something i know how to do yet and falls under the decoration phase which comes after the implementation of the rooms and major progression stuff. So I'll get there eventually.