Rafter entrance upper and lower finished except decoration, actual rafters in progress
Also did some work with the stairs, getting a weird bug that's documented in the stairs effect file. I suspect it's something to do with a way the player controller and the stairs effect interact, but because the stairs effect actually moves the player and doesn't apply any velocity there is no reason the player should keep having the effect when they leave the zone, and yet it happens anyway. Sometimes the effect is also reverse. It's very strange.